Fix net/ranking on broken profiles
Primary issue was multi-word last names—the two data sources (theunitedstates.io vs OpenSecrets) categorized the name parts differently. Solved by getting the last word of the last name.

Then came names with capitalized letters in the middle—so had to lower-case both, in addition to removing accents, standardizing quotes, etc.

Finally, two exception to the rule was names starting with "De" & "Des," of which there are 2—one source had a space after these, the other did not. I wrote a custom regex for this exception, which is insane. Look so simple on the outside!

The remaining issue is a few profiles not receiving rankings.